+++ examples/hybrid/staggered_nonhydrostatic_model.jl
@@ -59,10 +59,6 @@ const ᶠcurlᵥ = Operators.CurlC2F(
     bottom = Operators.SetCurl(CT12(FT(0), FT(0))),
     top = Operators.SetCurl(CT12(FT(0), FT(0))),
 )
-const ᶜFC = Operators.FluxCorrectionC2C(
-    bottom = Operators.Extrapolate(),
-    top = Operators.Extrapolate(),
-)
 const ᶠupwind_product1 = Operators.UpwindBiasedProductC2F()
 const ᶠupwind_product3 = Operators.Upwind3rdOrderBiasedProductC2F(
     bottom = Operators.ThirdOrderOneSided(),
@@ -164,10 +160,6 @@ function implicit_tendency!(Yₜ, Y, p, t)
     Yₜ.c.uₕ .= (zero(eltype(Yₜ.c.uₕ)),)
 
     @. Yₜ.f.w = -(ᶠgradᵥ(ᶜp) / ᶠinterp(ᶜρ) + ᶠgradᵥ(ᶜK + ᶜΦ))
-
-    # TODO: Add flux correction to the Jacobian
-    # @. Yₜ.c.ρ += ᶜFC(ᶠw, ᶜρ)
-    # @. Yₜ.c.ρe += ᶜFC(ᶠw, ᶜρe)
 
     return Yₜ
 end